The `setRoutes()` method is now available in `ConsoleApplication` class
